+++ spec/linter/color_keyword_spec.rb
@@ -21,6 +21,16 @@ describe SCSSLint::Linter::ColorKeyword do
     it { should report_lint line: 2 }
   end
 
+  context 'when a color keyword exists in a shorthand property' do
+    let(:css) { <<-CSS }
+      p {
+        border: 1px solid black;
+      }
+    CSS
+
+    it { should report_lint line: 2 }
+  end
+
   context 'when a property contains a color keyword as a string' do
     let(:css) { <<-CSS }
       p {
@@ -29,5 +39,25 @@ describe SCSSLint::Linter::ColorKeyword do
     CSS
 
     it { should_not report_lint }
+  end
+
+  context 'when a function call contains a color keyword' do
+    let(:css) { <<-CSS }
+      p {
+        color: function(red);
+      }
+    CSS
+
+    it { should report_lint line: 2 }
+  end
+
+  context 'when a mixin include contains a color keyword' do
+    let(:css) { <<-CSS }
+      p {
+        @include some-mixin(red);
+      }
+    CSS
+
+    it { should report_lint line: 2 }
   end
 end